requirements.txt
psutil
matplotlib
google-cloud-compute